Removing NoProject render in DocumentsList, doesnt belong here
