--- web/src/client/js/components/DocumentsList/NoProject.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const NoProject = () => {
-  return (
-    <div>
-      <h2>Project not found</h2>
-    </div>
-  )
-}
-
-export default NoProject
